requirements.txt
hyperliquid-python-sdk~=0.1.19